% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pdb.R
\name{cleaner}
\alias{cleaner}
\title{Organiza o PDB}
\usage{
cleaner(path_pdb)
}
\arguments{
\item{path_pdb}{One string that describe the walk to find the PDB file}
}
\description{
The code extracts only the 'ATOM' records
from a PDB file and removes hydrogens and most alternate
conformations.
}
\author{
Herson Hebert
}
