% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osfil.R
\name{create_osfil}
\alias{create_osfil}
\title{Construction of os.fil file.}
\usage{
create_osfil(first, last, raydist)
}
\arguments{
\item{first}{Number of first atom}

\item{last}{Number of last atom}

\item{raydist}{Character "y" or "n" that describes if the user wants create raydist}
}
\description{
The code is applied to buil os.fil file that is responsible to
store importants parameters to execution.
}
\author{
Herson Hebert
}
