% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_area.R
\name{ocluded_area}
\alias{ocluded_area}
\title{Calculating Occluded Surface Areas.}
\usage{
ocluded_area(path_pdb, first, last, raydist)
}
\arguments{
\item{path_pdb}{Name or path of pdb.}

\item{first}{Number of first atom.}

\item{last}{Number of last atom.}

\item{raydist}{Character that define if the user wants to generate raydist file.}
}
\description{
This a principal function of package osvfibo. It's a innovation
method based in Occluded Surface, but now using Fibonnaci Distribution.
}
\author{
Herson Hebert
}
