% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_read.R
\name{read_prot}
\alias{read_prot}
\title{Reading Values of Occluded Surface Area.}
\usage{
read_prot(prot)
}
\arguments{
\item{prot}{Name or path of prot.srf.}
}
\description{
It's a important function to load calculated values to use of user.
}
\author{
Herson Hebert
}
